(mostly) Working video tier!

+++ interfaces/chip8.go
@@ -15,7 +15,7 @@ type Chip8 struct {
 	DelayTimer byte
 	SoundTimer byte
 	Keypad     [16]byte
-	Video      [int(VIDEO_WIDTH) * int(VIDEO_HEIGHT)]uint32
+	Video      [int(VIDEO_WIDTH) * int(VIDEO_HEIGHT)]byte
 	Opcode     uint16
 	opcodes
 }
